% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLEAN.R
\name{clean}
\alias{clean}
\title{Apply CLEAN pipeline (minimal)}
\usage{
clean(df, id, time, y, contagion, A_by_time, method = "sbm")
}
\arguments{
\item{df}{panel dataframe}

\item{id}{node id column}

\item{time}{time column}

\item{y}{outcome column}

\item{contagion}{contagion regressor column (already constructed by user)}

\item{A_by_time}{list of adjacency matrices}

\item{method}{block method}
}
\value{
list with model + diagnostics
}
\description{
Apply CLEAN pipeline (minimal)
}
