% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLEAN.R
\name{clean_blocks_by_time}
\alias{clean_blocks_by_time}
\title{Detect blocks by time slice}
\usage{
clean_blocks_by_time(df, id, time, A_by_time, method = "sbm")
}
\arguments{
\item{df}{panel dataframe}

\item{id}{node id column name}

\item{time}{time column name}

\item{A_by_time}{named list of adjacency matrices}

\item{method}{block method}
}
\value{
df with a `clean_block` column
}
\description{
Detect blocks by time slice
}
