% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_datetime.R
\name{format_datetime}
\alias{format_datetime}
\title{Readable date format to POSIX}
\usage{
format_datetime(format)
}
\arguments{
\item{format}{the format that needs to be transformed}
}
\description{
Use this function to transform generic date/time info writing (dd-mm-yyyy) to POSIX standardised format (\%d/\%m/\%Y), see Examples.
}
\examples{
format_datetime("yyyy/mm/dd")
#> "\%Y/\%m/\%d"

# Very hard to remember all these characters:
format(Sys.time(), "\%a \%b \%d \%Y \%X")
#> "Fri Jul 12 2019 06:00:17"

# Easy to remember and write:
format(Sys.time(), format_datetime("ddd mmm dd yyyy hh:mm:ss"))
#> "Fri Jul 12 2019 06:00:17"
}
