\name{geneSigInClustering}
\alias{geneSigInClustering}
\alias{primSetSigInClustering}
\alias{geneSigInList}
\alias{getAllClusters}
\alias{pruneMinFDR}
\alias{insertCdt}
\alias{initializeCategories}
\alias{significantCategories}
\alias{geneSigInClustering_nullDistr}
\title{ Auxiliary functions to funcClustAnnot and geneListEnrichment }
\description{
  These functions are called by the generateTreeViewFiles(), runCLEAN() and the geneListEnrichment() functions.
}
\usage{
}
\details{
}
\value{
}
\author{ Johannes Freudenberg, Xiangdong Liu, Mario Medvedovic}
\note{

}
\seealso{ \code{\link{runCLEAN}}}
\examples{
}
\keyword{cluster}

