\name{plotCLEANscore}
\alias{plotCLEANscore}
\title{Function to generate a diagnostic plot using the CLEAN score}
\description{
    This function generates a plot of the number of genes with CLEANscores <= a threshold against that threshold.  These plots can be used, for example, to compare different clustering algorithms. 
}
\usage{
plotCLEANscore(fClustAnnotations, fCategoryName = "GO", maxCLEANscore = NULL, add = FALSE, ...)
}
\arguments{
  \item{fClustAnnotations}{A list of functional cluster annotations generated by the runCLEAN() function or the generateTreeViewFiles() function}
  \item{fCategoryName}{Name of the list of functional categories to be used for the comparison}
  \item{maxCLEANscore}{Maximal CLEANscore.  If specified, this parameter is used to specify the xlim parameter in the plot() function.}
  \item{add}{If TRUE, plot will be added to existing graph.}
  \item{...}{Additional parameters to the plot() function}
}
\author{Johannes Freudenberg}
\seealso{\code{\link{generateTreeViewFiles}}, \code{\link{runCLEAN}},\code{\link{plot}}, \code{\link{lines}}}
\examples{
## not run
#data(gimmOut)
#require(org.Rn.eg.db)
#allGenes <- unique(keys(org.Rn.egSYMBOL))
# plotCLEANscore is intended for larger datasets. Here we use the larger 
# background gene list to somewhat better demonstrate the 
# plotCLEANscore() function.  In real life, the background list should be
# the list of genes represented on the microarray
#fClustAnnotations <- runCLEAN(gimmOut, bkgList = allGenes, 
#	functionalCategories="CpGislands",species = "Rn", 
#	maxGenesInCategory=10000)
#plotCLEANscore(fClustAnnotations, fCategoryName="CpGislands")
}
\keyword{ aplot }
\keyword{ hplot }
\keyword{ cluster }

