% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unclean}
\alias{unclean}
\title{Example data that is not clean}
\format{A \code{\link{data.frame}} with 500 observations and the following variables:
\describe{
  \item{\code{date}}{Dates imported from Excel, they are integers ranging from ~30,000 to ~43,000.}
  \item{\code{gender}}{Characters with mixed values observed in original data about patients gender.}
}}
\usage{
unclean
}
\description{
This typical data example can be used for checking and cleaning.
}
\seealso{
\code{\link{freq}} to check values and \code{\link{clean}} to clean them.
}
\keyword{datasets}
